3rd Commit
Make some changes to userDetail controller, to add the option to delete
and update an existing user.

Changes to the database were made to add some information to user detail
such as position and experience, and later changing the form, to add
this extra features.

The home page wasnt displaying the user info, so I added the EL
necessary to do this.